Add page 5 and more template helpers
