+++ locales/en.json
@@ -24,6 +24,8 @@
   "errors.firstname.length": "Please enter your name",
   "errors.email.length": "Please enter a valid email address",
   "errors.description.length": "Please add a valid description",
+  "errors.exception.length": "Please add a valid exception",
+  "errors.supplier.length": "Please add a valid provider",
   "footer.privacy": "https://numerique.canada.ca/transparence/confidentialite/",
   "footer.privacy.desc": "Privacy policy",
   "footer.terms": "https://numerique.canada.ca/transparence/avis/",
@@ -66,21 +68,41 @@
   "Submit": "Submit",
   "wordmark.alt": "wordmark.alt",
   "ssc-tech.title": "Technical Exception",
-  "ssc-tech.intro": "Any urgent requests related to COVID-19 crisis management planning can be submitted as follows.",
+  "ssc-tech.intro": "Any urgent requests related to COVID-19 crisis management planning can be submitted as follows: Printers and Scanners: submit the order as a Time Sensitive Exception, email SSC.provisioningservice serviceapprovisionnement.SPC and include COVID-19 in the Subject header. Microcomputers: submit the order as a Time Sensitive Exception, email SSC.provisioningservice serviceapprovisionnement.SPC and include COVID-19 in the Subject header. Software: email SSC.provisioningservice-serviceapprovisionnement.SPC@canada.ca and include COVID-19 in the subject header. ",
   "ssc-tech.questions.description": "Submit the order as a Time Sensitive Exception, email SSC.provisioningservice-serviceapprovisionnement.SPC@canada.ca and include COVID-19 in the subject header.",
   "ssc-tech.order.title": "Order Summary",
+  "ssc-tech.protected": "Protected",
+  "ssc-tech.protected.1": "Protected A",
+  "ssc-tech.protected.2": "Protected B",
+  "ssc-tech.protected.3": "Protected C",
+  "ssc-tech.classified": "Classified",
+  "ssc-tech.classified.1": "Confidential",
+  "ssc-tech.classified.2": "Secret",
+  "ssc-tech.classified.3": "Top Secret",
+  "ssc-tech.nato": "NATO",
+  "ssc-tech.nato.1": "NATO Unclassified",
+  "ssc-tech.nato.2": "NATO Restricted",
+  "ssc-tech.nato.3": "NATO Confidential",
+  "ssc-tech.nato.4": "NATO Secret",
+  "ssc-tech.nato.5": "COSMIC Top Secret",
+  "ssc-tech.comsec": "COMSEC",
   "ssc-tech.questions.1": "Exception reason",
   "ssc-tech.questions.2": "Preferred provider",
   "ssc-tech.questions.3": "Request delivery date",
   "ssc-tech.questions.4": "Product quantity",
   "ssc-tech.questions.5": "Total order estimated cost",
   "ssc-tech-2.title": "Part A: Enter your contract information",
+  "ssc-tech-2.subtitle-1": "Indicate the type of access required",
+  "ssc-tech-2.subtitle-2": "Indicate the type of information that the supplier will be required to access",
+  "ssc-tech-2.subtitle-3": "Release Restrictions",
+  "ssc-tech-2.subtitle-4": "Level of information",
   "ssc-tech-2.intro": "Description goes here",
   "ssc-tech-2.questions.1": "Originating Government Department or Organization",
   "ssc-tech-2.questions.2": "Branch or Directorate",
   "ssc-tech-2.questions.3": "Subcontract Number",
   "ssc-tech-2.questions.4": "Name and Address of Subcontractor",
   "ssc-tech-2.questions.5": "Brief Description of Work",
+  "ssc-tech-2.questions.country": "Specify country",
   "ssc-tech-2.radio-1": "Will the supplier require access to Controlled Goods?",
   "ssc-tech-2.radio-2": "Will the supplier require access to unclassified military technical data subject to the provisions of the Technical Data Control Regulations?",
   "ssc-tech-2.radio-3": "Will the supplier and its employees require access to PROTECTED and/or CLASSIFIED information or assets?",
@@ -89,23 +111,21 @@
   "ssc-tech-3.title": "Part B: Enter personnel information",
   "ssc-tech-3.intro": "Description goes here",
   "ssc-tech-4.title": "Part C: Enter your safeguard (supplier) information",
-  "ssc-tech-5.title": "Part D: Enter your authorization information",
   "ssc-tech-4.intro": "Description goes here",
+  "ssc-tech-5.title": "Part D: Enter your authorization information",
+  "ssc-tech-5.intro": "Description goes here",
+  "ssc-tech.contact.name": "Full name",
+  "ssc-tech.contact.title": "Title",
+  "ssc-tech.contact.phone": "Telephone number",
+  "ssc-tech.contact.email": "E-mail address",
+  "ssc-tech.contact.fax": "Facsimilie number",
+  "ssc-tech.contact.date": "Date",
   "start_over": "start_over",
-  "start.enable_cookies_title": "start.enable_cookies_title",
-  "start.enable_cookies_body": "start.enable_cookies_body",
-  "Exception reason": "Exception reason",
-  "Preferred provider": "Preferred provider",
-  "Request delivery date": "Request delivery date",
-  "Product quantity": "Product quantity",
-  "Total order estimated cost": "Total order estimated cost",
-  "required": "required",
-  "ssc-tech-2.radio.1": "ssc-tech-2.radio.1",
-  "ssc-tech-2.radio.2": "ssc-tech-2.radio.2",
-  "ssc-tech-2.radio.3": "ssc-tech-2.radio.3",
-  "ssc-tech-2.radio.4": "ssc-tech-2.radio.4",
-  "ssc-tech-2.radio.5": "ssc-tech-2.radio.5",
-  "ssc-tech-2.radio.none": "ssc-tech-2.radio.none",
-  "ssc-tech-2.radio": "ssc-tech-2.radio",
-  "errors.name.length": "errors.name.length"
+  "Page 5 value": "Page 5 value",
+  "Full name": "Full name",
+  "Title": "Title",
+  "Telephone number": "Telephone number",
+  "E-mail address": "E-mail address",
+  "Facsimilie number": "Facsimilie number",
+  "Date": "Date"
 }
