+++ routes/ssc-tech-5/js/ssc-tech-5.js
@@ -1,0 +1,1 @@
+// your js code here
