+++ routes/ssc-tech-5/schema.js
@@ -1,0 +1,7 @@
+/* istanbul ignore file */
+
+const Schema = {}
+
+module.exports = {
+  Schema,
+}
