+++ views/base.njk
@@ -9,6 +9,8 @@
 {%- from 'input-file.njk' import fileInput as fileInput with context -%}
 {%- from 'form.njk' import form as form with context -%}
 {%- from 'banner.njk' import banner as banner with context -%}
+{%- from 'protected.njk' import protectedCheck as protectedCheck with context -%}
+{%- from 'contactInfo.njk' import contactInfo as contactInfo with context -%}
 
 <!DOCTYPE html>
 <html lang="{{ getLocale() }}">
